extract performPipelineLoad into its own file
